requirements.txt
colorama==0.4.6
python_magic==0.4.27
